Fix off by one error in Sophocles
The last sentence in the XML for Sophocles, *Oedipus Tyrannus* has `id="931"`,
but the Treebank Template configuration goes up to 932.

This PR changes the maximum to 931 so that the application doesn't error
when navigating through *Oedipus Tyrannus*.

A few of the other treebanks have a maximum chunk less than the XML. I'm
not sure if this is intended or not, so I didn't touch them. But it
might be worth taking a look!